pyproject.toml
[project]
name = "secret-bot"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Brian McCallister", email = "brianm@skife.org" }
]
dependencies = [
    "discord.py>=2.3.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
hello = "secret_bot:hello"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/secret_bot"]
